Add Item Factory and Drop Command
I've upgraded to an item factory. I know there's still a little refactoring to do, but I've added the item factory so that my code will be more extensible in the future. It's a little more than the MVP calls for at this stage, but better to refactor now than when the code is more complex