Move namespace of old code
I've changed the namespace of the old code to prevent overlap of naming. Since the basic idea of the project hasn't changed, there will be overlap of the old code naming and the new code. By changing the namespace I'll be able to have the names overlap without raising conflicts.