Chore: In progress Add CombatScene
This is still a work in progress, but I want to push before I go to sleep, just in case. I've got most of the CombatScene logic in place as well as the TODOs so I don't forget what I'm working on. Once I finish the CombatScene I'll be one step closer to having a cool, working, RPG