Feat: Add Quit Game Action
I was getting tired of not having an in-game way to end the game, so I added a quit game action to the player character. Now that the Player implements IActionProvider I can add any "always available" actions to the player. When I implement the save feature, I'll put it here as well.