Refactor menu system abstract and decouple (#1)
* Feat: Add ViewRenderer

This update implements the ViewRenderer, it's an improvement on the ConsoleRenderer and Menu. It abstracts things out to make it more testable and makes it more re-usable for different types of actions and environments.

* Fix: ConsoleViewRenderer repeating bug

The ConsoleViewRenderer was repeating the entire console content rather than simply redrawing the menu. This fixes that bug and redraws just the menu.

* Feat: Implement MenuService in CombatScene

I've added the newly made MenuService class in the CombatScene. This makes things more consistent and hopefully makes things easier to maintain long term.

* Chore: clear old code and update CombatScene

The old implementation has been removed and I had one more thing to update in the CombatScene. Now the new Menu system is the only thing I'm using throughout the game.